fix: Split utils/state.ts into utils/serialize-state.ts & utils/deserialize-state.ts. Closes #118 - Split utils/state.ts into utils/serialize-state.ts & utils/deserialize-state.ts - Use utils/deserialize-state.ts in entry-client.ts - Use utils/serialize-state.ts in entry-server.ts
