+++ src/utils/deserialize-state.ts
@@ -1,0 +1,8 @@
+export function deserializeState(state: string) {
+  try {
+    return JSON.parse(state || '{}')
+  } catch (error) {
+    console.error('[SSR] On state deserialization -', error, state)
+    return {}
+  }
+}
