+++ src/utils/serialize-state.ts
@@ -34,12 +34,3 @@ export function serializeState(state: any) {
     return '{}'
   }
 }
-
-export function deserializeState(state: string) {
-  try {
-    return JSON.parse(state || '{}')
-  } catch (error) {
-    console.error('[SSR] On state deserialization -', error, state)
-    return {}
-  }
-}
