+++ test/specs/utils/state.spec.ts
@@ -1,6 +1,6 @@
 import { test } from 'uvu'
 import * as assert from 'uvu/assert'
-import { serializeState } from '../../../src/utils/state'
+import { serializeState } from '../../../src/utils/serialize-state'
 
 test('serializeState', () => {
   // Simple case.
